requirements.txt
streamlit
requests
pinecone-client[grpc]
langchain
bcrypt
